+++ sw.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = "two-digit-addition-v7";
+const CACHE_NAME = "two-digit-addition-v8";
 const APP_FILES = [
   "./",
   "./index.html",